Rename illumination correction functions script.
